TradingProject/requirements.txt
django-admin==4.1.6
Python==3.10.9